+++ src/main/java/kitchenpos/eatinorders/domain/DeliveryAddress.java
@@ -26,4 +26,19 @@ public class DeliveryAddress {
     public String getValue() {
         return value;
     }
+
+    @Override public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        DeliveryAddress that = (DeliveryAddress) o;
+        return Objects.equals(value, that.value);
+    }
+
+    @Override public int hashCode() {
+        return Objects.hash(value);
+    }
 }
